+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/Option.java
@@ -252,6 +252,15 @@ public enum Option {
      */
     PLAIN_DEFINITION_KEYS(null, null),
     /**
+     * For the "format" attribute, JSON Schema defines various supported values. The OpenAPI specification assigns a few more of those in order to
+     * differentiate between standard data types (e.g. float vs. double) and even some more fixed data types (e.g. LocalDate, LocalDateTime) if
+     * {@link #ADDITIONAL_FIXED_TYPES} is enabled. By enabling this option, these extra "format" values will be set for the respective
+     * primitive/standard data types automatically.
+     *
+     * @since 4.15.0
+     */
+    EXTRA_OPEN_API_FORMAT_VALUES(null, null),
+    /**
      * Whether as the last step of the schema generation, unnecessary "allOf" elements (i.e. where there are no conflicts/overlaps between the
      * contained sub-schemas) should be merged into one, in order to make the generated schema more readable. This also applies to manually added
      * "allOf" elements, e.g. through custom definitions or attribute overrides.